[SYCL][E2E] Split WG Memory basic usage test
